+++ src/message.c
@@ -1389,6 +1389,8 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
                 daemon_fail(rsp, "cannot focus space because the display is in the middle of an animation.\n");
             } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
                 daemon_fail(rsp, "cannot focus space because mission-control is active.\n");
+            } else {
+                daemon_fail(rsp, "cannot focus space due to an error with the scripting-addition.\n");
             }
         }
     } else if (token_equals(command, COMMAND_SPACE_MOVE)) {