+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 ### Changed
 - Update scripting-addition to support macOS Big Sur 11.0 Build 20A5384c [#589](https://github.com/koekeishiya/yabai/issues/589)
+- Return a non-zero exit code when focusing a space fails due to an issue with the scripting-addition [#181](https://github.com/koekeishiya/yabai/issues/181)
 
 ## [3.3.0] - 2020-09-03
 ### Added